+++ modules/nlu/src/backend/engine2/entity-extractor.ts
@@ -3,9 +3,8 @@ import _ from 'lodash'
 import jaroDistance from '../tools/jaro'
 import levenDistance from '../tools/levenshtein'
 import { extractPattern } from '../tools/patterns-utils'
-import { EntityExtractionResult, ListEntityModel, PatternEntity } from '../typings'
+import { EntityExtractionResult, ListEntityModel, PatternEntity, Tools } from '../typings'
 
-import { Tools } from './engine2'
 import { Predictors } from './predict-pipeline'
 import { TrainOutput } from './training-pipeline'
 import Utterance, { UtteranceToken } from './utterance'
