+++ modules/nlu/src/backend/engine2/training-pipeline.ts
@@ -4,13 +4,12 @@ import _ from 'lodash'
 import tfidf from '../pipelines/intents/tfidf'
 import { replaceConsecutiveSpaces } from '../tools/strings'
 import { isSpace, SPACE } from '../tools/token-utils'
-import { ListEntity, ListEntityModel, PatternEntity, TFIDF, Token2Vec, TrainingSession } from '../typings'
+import { ListEntity, ListEntityModel, PatternEntity, TFIDF, Token2Vec, Tools, TrainingSession } from '../typings'
 
 import CRFExtractor2 from './crf-extractor2'
-import { Tools } from './engine2'
 import { extractUtteranceEntities } from './entity-extractor'
 import { Model } from './model-service'
-import Utterance, { buildUtterances, UtteranceToken, UtteranceToStringOptions } from './utterance'
+import Utterance, { buildUtteranceBatch, UtteranceToken, UtteranceToStringOptions } from './utterance'
 
 // TODO make this return artefacts only and move the make model login in E2
 export type Trainer = (input: TrainInput, tools: Tools) => Promise<Model>
@@ -257,7 +256,7 @@ export const ProcessIntents = async (
 ): Promise<Intent<Utterance>[]> => {
   return Promise.map(intents, async intent => {
     const cleaned = intent.utterances.map(replaceConsecutiveSpaces)
-    const utterances = await buildUtterances(cleaned, languageCode, tools)
+    const utterances = await buildUtteranceBatch(cleaned, languageCode, tools)
 
     const allowedEntities = _.chain(intent.slot_definitions)
       .flatMap(s => s.entities)
@@ -312,7 +311,7 @@ export const AppendNoneIntents = async (input: TrainOutput, tools: Tools): Promi
   const intent: Intent<Utterance> = {
     name: NONE_INTENT,
     slot_definitions: [],
-    utterances: await buildUtterances(noneUtterances, input.languageCode, tools),
+    utterances: await buildUtteranceBatch(noneUtterances, input.languageCode, tools),
     contexts: [...input.contexts],
     vocab: {},
     slot_entities: []
